final code cleanup (at least until 0.8.1)
